Ready for release 2.1 of BRL CAD.
svn:revision:2473
cvs:account:moss
cvs:branch:trunk

+++ lgt/getfont.c
@@ -26,7 +26,7 @@ static char RCSid[] = "@(#)$Header$ (BRL)";
 
 /* Variables controlling the font itself */
 FILE		*ffdes;		/* File pointer for current font.	*/
-int		offset;		/* Current offset to character data.	*/
+long		offset;		/* Current offset to character data.	*/
 struct header	hdr;		/* Header for font file.		*/
 struct dispatch	dir[256];	/* Directory for character font.	*/
 int		width = 0,	/* Size of current character.		*/
